% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradconj.R
\name{gradconj}
\alias{gradconj}
\title{conjugate gradient algorithm}
\usage{
gradconj(A, b, x0, tol, max_iter, plot_chart = TRUE)
}
\arguments{
\item{A}{positive definite symmetric matrix}

\item{b}{a vector}

\item{x0}{initial value}

\item{tol}{tolerance}

\item{max_iter}{maximum number of iterations}
}
\value{
solution Ax=b
}
\description{
conjugate gradient algorithm
}
\examples{
 A <- matrix(c(4, 1, 1, 3), nrow = 2)
b <- c(1, 2)
x0 <- c(0, 0)
tol <- 1e-10
maxiter <- 1000
gradconj(A, b, x0, tol, maxiter)
}
