% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradconj.R
\name{YANKAM}
\alias{YANKAM}
\title{algorithme du gradient conjugué}
\usage{
YANKAM(A, b, x0, tol, max_iter)
}
\arguments{
\item{A}{une matrice definie positive et symetrique}

\item{b}{un vecteur}

\item{x0}{valeur un intiale}

\item{tol}{tolerance}

\item{max_iter}{nombre maximum d'ittération}
}
\value{
la solution de Ax=b
}
\description{
algorithme du gradient conjugué
}
\examples{
 A <- matrix(c(4, 1, 1, 3), nrow = 2)
b <- c(1, 2)
x0 <- c(0, 0)
tol <- 1e-10
maxiter <- 1000
YANKAM(A, b, x0, tol, maxiter)
}
